% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{get_K_cos}
\alias{get_K_cos}
\title{get_K_cos.}
\usage{
get_K_cos(sq_dist, phi)
}
\arguments{
\item{sq_dist}{is a matrix of Euclidean distance of all spots}

\item{phi}{is a periodicity parameter of the Cosine kernel}
}
\value{
a cosine kernel
}
\description{
get_K_cos used to get the Cosine kernel.
}
\details{
get_K_cos used to get the Cosine kernel.
}
