% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{get_K_gauss}
\alias{get_K_gauss}
\title{get_K_gauss.}
\usage{
get_K_gauss(sq_dist, length_scale)
}
\arguments{
\item{sq_dist}{is a matrix of Euclidean distance of all spots}

\item{length_scale}{is a length scale parameter of the Gaussian kernel}
}
\value{
a Gaussian kernel
}
\description{
get_K_gauss used to get the Gaussian kernel.
}
\details{
get_K_cos used to get the Gaussian kernel.
}
