% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{get_sq_dist}
\alias{get_sq_dist}
\title{get_sq_dist.}
\usage{
get_sq_dist(spa_mat)
}
\arguments{
\item{spa_mat}{is a matrix of Euclidean distance of all spots}
}
\value{
a matrix of Euclidean distance
}
\description{
get_sq_dist used to get Euclidean distance of all spots.
}
\details{
get_sq_dist used to get Euclidean distance of all spots.
}
