% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{select_gene}
\alias{select_gene}
\title{select_gene.}
\usage{
select_gene(fit)
}
\arguments{
\item{fit}{a list}
}
\value{
a list
}
\description{
select_gene used to select the genes that have the same type of kernels under both null and alternative hypothesis.
}
\details{
select_gene used to select the genes that have the same type of kernels under both null and alternative hypothesis.
}
