% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatialDEG}
\alias{spatialDEG}
\title{spatialDEG.}
\usage{
spatialDEG(
  Y1,
  Y2,
  spa1 = NULL,
  spa2 = NULL,
  W1,
  W2,
  Initial_theta = matrix(1, 3, 1),
  num_ls = 10,
  Kernel1 = NULL,
  Kernel2 = NULL,
  parallel = TRUE,
  kernel_fixed = FALSE,
  kernel_matched = FALSE,
  coreNum = 1,
  max_Iter = 1000,
  check_positive = TRUE,
  eps = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{Y1}{is a vector of gene expression for the first dataset.}

\item{Y2}{is a vector of gene expression for the second dataset.}

\item{spa1}{is a matrix of spatial location for the first dataset.}

\item{spa2}{is a matrix of spatial location for the second dataset.}

\item{W1}{is a matrix of covariates for the first dataset.}

\item{W2}{is a matrix of covariates for the second dataset.}

\item{Initial_theta}{is a vector of initial value for theta. The default vector is ones}

\item{num_ls}{is a integer specifying the number of length scale or periodicity parameters.}

\item{Kernel1}{is a matrix of Kernel for the first dataset.}

\item{Kernel2}{is a matrix of Kernel for the second dataset.}

\item{parallel}{is a logical value specifying whether to run spatialDEG in parallel.}

\item{kernel_fixed}{Kernel_fixd a logical value specifying whether the kernel is fixed}

\item{kernel_matched}{a logical value specifying whether the the kernels under both null hypothesis and alternative hypothesis are restricted to be the same type of kernel}

\item{coreNum}{is an integer specifying the number of cores used for parallel computation.}

\item{max_Iter}{is an integer specifying the maximum iteration of the AI algorithm.}

\item{check_positive}{a logical value specifying whether to check the positive definiteness of kernels}

\item{eps}{a value specifying when the AI average convergence.}

\item{verbose}{is logical value specifying whether to print the parameters during spatialDEG analysis}
}
\value{
a list
}
\description{
spatialDEG used to perform differentially expressed gene analysis by leveragin spatial informaiton.
}
\details{
spatialDEG used to perform differentially expressed gene analysis by leveragin spatial informaiton.
}
