% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatialDEG_paral_test}
\alias{spatialDEG_paral_test}
\title{spatialDEG_paral_test}
\usage{
spatialDEG_paral_test(
  spa1,
  spa2,
  W,
  Y,
  Initial_theta,
  num_ls = 20L,
  iteration = 100L,
  Kernel_fixd = FALSE,
  verbose = FALSE,
  kernel_matched = FALSE,
  coreNum = 1L,
  tol = 1e-04
)
}
\arguments{
\item{spa1}{spatial location for the first dataset.}

\item{spa2}{spatial location for the second dataset.}

\item{W}{covariates for the two datasets}

\item{Y}{stacked gene expression}

\item{Initial_theta}{initial value of theta}

\item{num_ls}{total number of length scale and periodicity}

\item{iteration}{The maximum iteration}

\item{Kernel_fixd}{a logical value specifying whether the kernel is fixed}

\item{verbose}{a logical value specifying whether to print the parameters during the spatialDEG analysis or not}

\item{kernel_matched}{a logical value specifying whether the the kernels under both null hypothesis and alternative hypothesis are restricted to be the same type of kernel}

\item{coreNum}{Number of cores used for parallel computation}

\item{tol}{tolence to stop the algorithm}
}
\value{
List of model parameters
}
\description{
spatialDEG to perform differentially expressed gene analysis by leveraging spatial information in parallel.
}
\details{
\code{spatialDEG_paral_test} run spatialDEG model.
}
\author{
Yi Yang
}
