% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatialDEG_test}
\alias{spatialDEG_test}
\title{spatialDEG}
\usage{
spatialDEG_test(
  spa1,
  spa2,
  W,
  Y,
  Initial_theta,
  num_ls = 20L,
  mu_fixed = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spa1}{spatial location for the first dataset.}

\item{spa2}{spatial location for the second dataset.}

\item{W}{covariates for the two datasets}

\item{Y}{stacked gene expression}

\item{Initial_theta}{initial value of theta}

\item{num_ls}{total number of length scale and periodicity}

\item{mu_fixed}{logical value specifying whether mu is fixed}

\item{verbose}{logical value specifying whether to print the parameters during the spatialDEG analysis or not}
}
\value{
List of model parameters
}
\description{
spatialDEG to perform differentially expressed gene analysis by leveraging spatial information.
}
\details{
\code{spatialDEG_test} run spatialDEG model.
}
\author{
Yi Yang
}
