% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatialDEG_true_kernel_test}
\alias{spatialDEG_true_kernel_test}
\title{spatialDEG}
\usage{
spatialDEG_true_kernel_test(
  W,
  Y,
  K1,
  K2,
  Initial_theta,
  mu_fixed = TRUE,
  verbose = TRUE,
  check_positive = FALSE
)
}
\arguments{
\item{W}{covariates for the two datasets}

\item{Y}{stacked gene expression}

\item{K1}{kernel for the first dataset.}

\item{K2}{kernel for the second dataset.}

\item{Initial_theta}{initial value of theta}

\item{mu_fixed}{logical value specifying whether mu is fixed}

\item{verbose}{logical value specifying whether to print the parameters during the spatialDEG analysis or not}

\item{check_positive}{a logical value specifying whether to check the positive definiteness of kernels}
}
\value{
List of model parameters
}
\description{
spatialDEG to perform differentially expressed gene analysis by leveraging spatial information with true kernels.
}
\details{
\code{spatialDEG_test} run spatialDEG model.
}
\author{
Yi Yang
}
