% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SysMatEigen2}
\alias{SysMatEigen2}
\title{SysMatEigen2}
\usage{
SysMatEigen2(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
a List
}
\description{
SysMatEigen2 to transform the matrix to a positive definite matrix
}
\details{
\code{spatialDEG_test} run spatialDEG model.
}
\author{
Yi Yang
}
